requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0.1
django-appconf==1.0.6
django-compressor==4.4
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2023.4
rcssmin==1.1.1
reportlab==4.0.9
rjsmin==1.2.1
sqlparse==0.4.4
typing_extensions==4.9.0
